+++ examples/src/store/Store.ts
@@ -1,5 +1,4 @@
-import { LocationManager } from '@moxb/moxb';
-import { TokenManager, TokenManagerImpl } from '../../../packages/moxb/src/routing';
+import { LocationManager, TokenManager, TokenManagerImpl, LinkGenerator, LinkGeneratorImpl } from '@moxb/moxb';
 import { Application } from '../app/Application';
 import { ApplicationImpl } from '../app/ApplicationImpl';
 import { MemTable } from '../memtable/MemTable';
@@ -9,6 +8,7 @@ import { UrlStore } from './UrlStore';
 import { UrlStoreImpl } from './UrlStoreImpl';
 import { ViewStore } from './ViewStore';
 import { ViewStoreImpl } from './ViewStoreImpl';
+import { mainMenu } from '../MenuStructure';
 
 export interface Store {
     readonly app: Application;
@@ -17,6 +17,7 @@ export interface Store {
     readonly locationManager: LocationManager;
     readonly tokenManager: TokenManager;
     readonly url: UrlStore;
+    readonly linkGenerator: LinkGenerator;
 }
 
 export class StoreImpl implements Store {
@@ -26,6 +27,7 @@ export class StoreImpl implements Store {
     readonly locationManager: LocationManager;
     readonly tokenManager: TokenManager;
     readonly url: UrlStore;
+    readonly linkGenerator: LinkGenerator;
 
     constructor() {
         this.locationManager = new LocationStoreImpl();
@@ -34,5 +36,10 @@ export class StoreImpl implements Store {
         this.app = new ApplicationImpl();
         this.memTable = new MemTableImpl(this.url, this.tokenManager);
         this.view = new ViewStoreImpl();
+        this.linkGenerator = new LinkGeneratorImpl({
+            locationManager: this.locationManager,
+            states: mainMenu,
+            rootUrl: 'https://localhost:3100',
+        });
     }
 }